# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DomainTopLists": {
        "properties": {
          "top_domains": {
            "items": {
              "$ref": "#/components/schemas/DomainRow"
            },
            "title": "Top Domains",
            "type": "array"
          },
          "top_unfetchable": {
            "items": {
              "$ref": "#/components/schemas/DomainRow"
            },
            "title": "Top Unfetchable",
            "type": "array"
          }
        },
        "required": [
          "top_domains",
          "top_unfetchable"
        ],
        "title": "DomainTopLists",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first research library API",
    "title": "Alexandria",
    "version": "0.0.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ingestion/domains": {
      "get": {
        "description": "Top domains by document count and by unfetchable count.",
        "operationId": "domain_top_lists_ingestion_domains_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainTopLists"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Domain Top Lists",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/ingestion/enrich": {
      "post": {
        "description": "Re-run an enrichment pass over documents missing that artifact.\n\nThe retrigger for ``purge summaries``: its skip gate is \"has any chunk\",\nwhich a summary purge leaves true, so re-syncing would not regenerate it.",
        "operationId": "enrich_endpoint_ingestion_enrich_post",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "summary",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich Endpoint",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/ingestion/enrichment-runs": {
      "get": {
        "description": "What ran, when, with which model, and at what cost in provider traffic.\n\nThe provenance surface behind the purge filters below: a run listed here is\na `run_id` a purge can target.",
        "operationId": "enrichment_runs_list_ingestion_enrichment_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrichment Runs List",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/ingestion/purge-targets": {
      "get": {
        "description": "The purge registry with live dry-run counts \u2014 one row per button.",
        "operationId": "purge_targets_ingestion_purge_targets_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge Targets",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/ingestion/purge/{key}": {
      "post": {
        "description": "Purge one registered target, optionally scoped by source and provenance.\n\n``run_id`` / ``provider`` / ``model`` narrow to what a particular backend\nproduced; ``unknown=true`` selects the pre-provenance backlog. Targets whose\nartifact carries no run stamp reject those filters with a 400 rather than\nsilently widening the purge.",
        "operationId": "purge_target_endpoint_ingestion_purge__key__post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "unknown",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unknown",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge Target Endpoint",
        "tags": [
          "ingestion"
        ]
      }
    },
...
    "/ingestion/sources/{source}/browse": {
      "post": {
        "operationId": "browse_path_ingestion_sources__source__browse_post",
        "parameters": [
          {
            "in": "path",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Browse Path",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}